requirements.txt
setuptools~=67.7.2
requests~=2.29.0
pytest~=7.3.1
responses~=0.23.1
openapi-core~=0.17.0
PyYAML~=6.0
werkzeug~=2.3.3
Flask~=2.3.2
mariadb~=1.0.11
